Add support for recursive functions ...
... (and allow factoring `#[with]` annotations within a `trait` or `impl`).

This is achieved by using `::func_wrap` to always forward the generic
implementation to one that uses a fixed type (`dyn FnMut`) which lets us
avoid the infinite type recursion error.

That being said, `::func_wrap`'s forwarding cannot refer to types
introduced by an `impl` or `trait` block unless such as block is, in and
on itself, `#[with]`-annotated (so as to forward that info to it).

This leads to a robust implementation (compared to the experimental
`recursive_fix` branch), so it will be merged.

That being said, it does bump the MSRV to 1.42.0.

In all cases, it will suppose a major semver bump (TODO).